+++ react/components/TooltipVerification.js
@@ -1,17 +1,21 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { injectIntl, intlShape } from 'react-intl'
-import { isMobile } from 'react-device-detect'
+
 import { translate } from '../utils/translate'
 import Tooltip from './Tooltip'
 import PasswordValidationContent from './PasswordValidationContent'
+import getIsMobile from '../utils/getIsMobile'
 
 class TooltipVerification extends Component {
   render() {
     const { fields, intl } = this.props
-    if (isMobile) {
+    if (getIsMobile()) {
       return (
-        <Tooltip top title={translate('store/login.password.tooltip.title', intl)}>
+        <Tooltip
+          top
+          title={translate('store/login.password.tooltip.title', intl)}
+        >
           <PasswordValidationContent fields={fields} />
         </Tooltip>
       )
@@ -28,7 +32,7 @@ class TooltipVerification extends Component {
 TooltipVerification.propTypes = {
   /** Fields to be verified in the tooltip */
   fields: PropTypes.array,
-  /** Intl object*/
+  /** Intl object */
   intl: intlShape,
 }