+++ react/utils/getIsMobile.js
@@ -1,0 +1,9 @@
+const getIsMobile = () => {
+  return (
+    window &&
+    window.matchMedia &&
+    window.matchMedia('(max-width: 40em)').matches
+  )
+}
+
+export default getIsMobile